refactor: PR 리뷰 반영 — 어노테이션, 도메인 분리, Transactional 메서드 단위
- @AuthAccountId 커스텀 어노테이션 + ArgumentResolver 추가
- UserInterestStore 포트를 user 도메인 패키지로 분리
- UserInterestStoreAdapter를 interest 패키지로 분리
- @Transactional 클래스 → 메서드 단위로 변경 (전체 UseCase)
