+++ src/main/kotlin/com/chat/chingudachi/application/auth/port/PersistencePorts.kt
@@ -4,8 +4,6 @@ import com.chat.chingudachi.domain.account.Account
 import com.chat.chingudachi.domain.account.AccountCredential
 import com.chat.chingudachi.domain.account.CredentialType
 import com.chat.chingudachi.domain.auth.AuthToken
-import com.chat.chingudachi.domain.interest.UserInterest
-
 interface AccountStore {
     fun save(account: Account): Account
     fun findById(id: Long): Account?
@@ -14,10 +12,6 @@ interface AccountStore {
 interface AccountCredentialStore {
     fun save(credential: AccountCredential): AccountCredential
     fun findByCredentialTypeAndOauthKey(credentialType: CredentialType, oauthKey: String): AccountCredential?
-}
-
-interface UserInterestStore {
-    fun findByAccountId(accountId: Long): List<UserInterest>
 }
 
 interface AuthTokenStore {
