+++ src/main/kotlin/com/chat/chingudachi/application/user/GetMyProfileUseCase.kt
@@ -1,7 +1,7 @@
 package com.chat.chingudachi.application.user
 
 import com.chat.chingudachi.application.auth.port.AccountStore
-import com.chat.chingudachi.application.auth.port.UserInterestStore
+import com.chat.chingudachi.application.user.port.UserInterestStore
 import com.chat.chingudachi.domain.account.Account
 import com.chat.chingudachi.domain.common.ErrorCode
 import com.chat.chingudachi.domain.common.NotFoundException
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
-@Transactional(readOnly = true)
 class GetMyProfileUseCase(
     private val accountStore: AccountStore,
     private val userInterestStore: UserInterestStore,
 ) {
+    @Transactional(readOnly = true)
     fun execute(accountId: Long): MyProfile {
         val account = accountStore.findById(accountId)
             ?: throw NotFoundException(ErrorCode.ACCOUNT_NOT_FOUND)
