+++ src/main/kotlin/com/chat/chingudachi/application/user/port/UserInterestStore.kt
@@ -1,0 +1,7 @@
+package com.chat.chingudachi.application.user.port
+
+import com.chat.chingudachi.domain.interest.UserInterest
+
+interface UserInterestStore {
+    fun findByAccountId(accountId: Long): List<UserInterest>
+}
