+++ src/main/kotlin/com/chat/chingudachi/presentation/common/AuthAccountId.kt
@@ -1,0 +1,26 @@
+package com.chat.chingudachi.presentation.common
+
+import com.chat.chingudachi.infrastructure.security.SecurityContextUtil
+import org.springframework.core.MethodParameter
+import org.springframework.stereotype.Component
+import org.springframework.web.bind.support.WebDataBinderFactory
+import org.springframework.web.context.request.NativeWebRequest
+import org.springframework.web.method.support.HandlerMethodArgumentResolver
+import org.springframework.web.method.support.ModelAndViewContainer
+
+@Target(AnnotationTarget.VALUE_PARAMETER)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class AuthAccountId
+
+@Component
+class AuthAccountIdArgumentResolver : HandlerMethodArgumentResolver {
+    override fun supportsParameter(parameter: MethodParameter): Boolean =
+        parameter.hasParameterAnnotation(AuthAccountId::class.java)
+
+    override fun resolveArgument(
+        parameter: MethodParameter,
+        mavContainer: ModelAndViewContainer?,
+        webRequest: NativeWebRequest,
+        binderFactory: WebDataBinderFactory?,
+    ): Long = SecurityContextUtil.getCurrentAccountId()
+}
